Containerize infoblox-daemon
Refactoring and code clean up:
- instead of having a copy of skel.go from CNI, we just
  extend skel.CmdArgs class to include IfMac (see cmdargs.go)
- Remove debug output
- Add license txt
- renamed a couple of filenames.
